Split-off C++11 sync and POSIX sync. Added conditionals for MAF. Added proper error handling for threads
