+++ srtcore/utilities.h
