+++ srtcore/window.h
@@ -148,6 +148,7 @@ public:
         m_tsProbeTime(),
         m_Probe1Sequence(SRT_SEQNO_NONE)
     {
+        // Exception: up to CUDT ctor
         srt::sync::setupMutex(m_lockPktWindow, "PktWindow");
         srt::sync::setupMutex(m_lockProbeWindow, "ProbeWindow");
         CPktTimeWindowTools::initializeWindowArrays(m_aPktWindow, m_aProbeWindow, m_aBytesWindow, ASIZE, PSIZE);
